requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2.31
fredapi>=0.5.1
requests>=2.31
pydantic>=2.0
pyyaml>=6.0
streamlit>=1.28
matplotlib>=3.7
quantstats>=0.0.62
pyarrow>=14.0
pytest>=7.4
pandas_market_calendars>=4.3
